changed reward to integer and cumlative, now counts number of rounds

+++ models/GameState.cs
@@ -8,7 +8,7 @@ namespace CardExploration.models
         ///<summery>
         ///Stores aspects of the game state including player hand, dealer hand and  current reward
         ///<summery>
-        double Reward {get; set;}
+        Int32 Reward {get; set;}
         long State {get; set;}
 
     }
